added exclusions to pf regex
