+++ cishouseholds/regex/healthcare_regex.py
@@ -1,18 +1,8 @@
 from collections import namedtuple
 
+from cishouseholds.regex import match_with_exclusions
+
 RegexPattern = namedtuple("RegexPattern", ["positive_regex_pattern", "negative_regex_pattern"])
-
-
-def match_with_exclusions(patters_to_match, patterns_to_exclude=[]):
-    if type(patters_to_match) != list:
-        patters_to_match = [patters_to_match]
-    if type(patterns_to_exclude) != list:
-        patterns_to_exclude = [patterns_to_exclude]
-    patterns_to_exclude = "|".join(patterns_to_exclude)
-    regex = ""
-    for pattern in patters_to_match:
-        regex += rf"(?=.*?({pattern}))"
-    return regex + rf"(^(?!.*({patterns_to_exclude})).*)"
 
 
 # exclusions
